debugging decoder

+++ src/fastertransformer/models/whisper/CachedSequenceAttention.cc
@@ -8,8 +8,8 @@ void CachedSequenceAttention<T>::forward(TensorMap output_tensors, TensorMap inp
 input_tensors:
     input_queries : [batch, q_seq, d_model]
     input_keys : [batch, k_seq, d_model]
-    input_lengths: size_t[batch] (optional, mask is autoregressive if none)
-    step : size_t[1]
+    input_lengths: uint32_t[batch] (optional, mask is autoregressive if none)
+    step : uint32_t[1]
     transpose_key_cache: (optional), transpose key_cache to batch, num_heads, size_per_head/x, seq, x if not null
 
 output_tensors:
@@ -20,10 +20,10 @@ output_tensors:
 {
     Tensor input_queries = input_tensors.at("input_queries");
     Tensor input_keys = input_tensors.at("input_keys");
-    size_t batch = input_queries.shape[0];
-    size_t q_seq = input_queries.shape[1];
-    size_t d_model = input_queries.shape[2];
-    size_t 
+    uint32_t batch = input_queries.shape[0];
+    uint32_t q_seq = input_queries.shape[1];
+    uint32_t d_model = input_queries.shape[2];
+    uint32_t 
 }
 
 template class CachedSequenceAttention<float>;
