+++ src/fastertransformer/th_op/whisper/VectorReader.h
@@ -5,7 +5,7 @@ namespace th = torch;
 namespace torch_ext{
 template <typename T> 
 struct VectorReader {
-    size_t idx;
+    uint32_t idx;
     std::vector<th::Tensor> *vec_;
     VectorReader(std::vector<th::Tensor> *vec): idx(0), vec_(vec) {};
     T* read() { th::Tensor t = vec_->at(idx++); return get_ptr<T>(t); }
